Mp4\seminars
git-svn-id: https://svn.esynctraining.com/svn/esynctraining@10296 41667c05-fd31-8a45-9ea7-b4ae18c5af78

+++ EdugameCloud.Lti.Core/Business/Models/LmsCourseMeetingModel.cs
@@ -31,7 +31,7 @@
 
         #endregion
         
-        public LmsCourseMeeting GetOneByCourseAndId(int companyLmsId, int courseId, int id)
+        public LmsCourseMeeting GetOneByCourseAndId(int companyLmsId, int courseId, long id)
         {
             if (companyLmsId <= 0)
                 throw new ArgumentOutOfRangeException("companyLmsId");
